Move default system clock impl to clock.rs
