+++ slatedb/src/config.rs
@@ -285,45 +285,6 @@ pub enum Ttl {
     ExpireAfter(u64),
 }
 
-/// defines the clock that SlateDB will use during this session
-pub trait Clock {
-    /// Returns a timestamp (typically measured in millis since the unix epoch),
-    /// must return monotonically increasing numbers (this is enforced
-    /// at runtime and will panic if the invariant is broken).
-    ///
-    /// Note that this clock does not need to return a number that
-    /// represents the unix timestamp; the only requirement is that
-    /// it represents a sequence that can attribute a logical ordering
-    /// to actions on the database.
-    fn now(&self) -> i64;
-}
-
-/// contains the default implementation of the Clock, and will return the system time
-#[derive(Default)]
-pub struct SystemClock {
-    last_tick: AtomicI64,
-}
-
-impl SystemClock {
-    pub fn new() -> Self {
-        Self {
-            last_tick: AtomicI64::new(i64::MIN),
-        }
-    }
-}
-
-impl Clock for SystemClock {
-    fn now(&self) -> i64 {
-        // since SystemTime is not guaranteed to be monotonic, we enforce it here
-        let tick = match SystemTime::now().duration_since(UNIX_EPOCH) {
-            Ok(duration) => duration.as_millis() as i64, // Time is after the epoch
-            Err(e) => -(e.duration().as_millis() as i64), // Time is before the epoch, return negative
-        };
-        self.last_tick.fetch_max(tick, SeqCst);
-        self.last_tick.load(SeqCst)
-    }
-}
-
 /// Defines the scope targeted by a given checkpoint. If set to All, then the checkpoint will
 /// include all writes that were issued at the time that create_checkpoint is called. If force_flush
 /// is true, then SlateDB will force the current wal, or memtable if wal_enabled is false, to flush
